Basket/21-12-2025-habit-tracker-bot/requirements.txt
python-telegram-bot[job-queue]>=22.5
python-dotenv
pytz
apscheduler
cachetools